User model Added read-only field 'admin' An user is the admin if and only if his email is equal to a hardcoded value.
In the future this might be replaced with a db backed field, or
something else.

Added the skeleton of UsersController responding to following actions:
- Index
- Create
- Update
- Destroy

Until authentication management is built, this controller is not
protected in any ways.

This allows us to develop user management interface both server and
client side.