Client Side
Added spinner handling code.
As of now, each and every HTTP request shows a spinner.
When the request terminates (doesn't matter whether successfully
or not), the spinner is hidden.

Login form is artificially hidden since we aren't dealing with
authentication yet.

Added:
UsersCtrl#index

This fetches the users from the backend and updates the $scope.users
accordingly.
The view shows a table with users fetched from the backend.