Client Side Fixed a bug in generic HTTP interceptor
Both in success and error cases of the interceptor the same
function was used to stop the spinner.
That function always returned the promise handed to it regardless
of whether the actual request succeeded or not.

Now two separate functions deal with error and success cases,
each one calling the stopSpinner function and then passing
either the response (success) or a rejection down the chain.

Server Side
Added uniqueness validation to user's email