Client Side Validation errors management
Each field in the user form has a class added to its control-group.
The class name must be equal to field name.

That way when a validation error is returned from backend
the UserCtrl can easily interact with each element of the input
form and mark fields with error and eventually be able to add
error messages for that field.

So far, everything has been tested during users creation and when
each field has only one error.