Client Side User management Editing or creating a new user now hides the users list and shows the editing form. The form gets hidden upon pressing the "Annulla" button. This is ok for the updating action, while it's a bit weird during creation phase. After creation the form should be hidden and the updated users list shown.
As of now any error reported from the backend during creation/update
of an user, is handed to the showValidationErrors function.
That function expects validation errors to be returned in hash format.

Server Side
UsersController

The controller's actions all take advantage of the respond_with
protocol in Rails.
It might be useful to add RABL to the application's gems requirement,
so that customization of the JSON response can take place.
This would simplify dealing with resources with relations.
Infact, responding with all user's RPS could leverage the template
from RPS and thus enabling composition/inheritance.
Plus it could be possible to customize reporting error conditions.