Server Side Added RABL gem to enhance JSON API
The gem allowed to specify complex JSON replies from the backend, as
well as composition and reuse.

It has been required the addition of a RablResponder to Rails.
That responder handles the case when the action has completed with
validation errors. Under that circumstance, it forces the rendering
of the RABL template, which could add detailed informations on the
errors and their nature.

Client Side
Only one small change has been made to UsersCtrl to handle the
case where an error_msg is provided from the backend.

Later on, it might be required to DRY-up error handling as controllers
start to proliferate.