Client Side Fixed a typo in SessionSvc
The typo prevented the rest of the code to properly determine if a user
was currently logged in or not.

In session service the currently logged in user was stored in
current_user
while the rest of the code looked for currentUser which was always
undefined ...

Now the variable is properly named as currentUser everywhere.

Login failure cases are yet to be tested. Logout is not implemented yet.