Server Side Default Responder changed to RablResponder
This change has application wide effects now.
ApplicationController's class method 'responder' has been overridden
so that it could return our custom responder.

Client Side
Fixed a bug that prevented validation errors to be shown on the
form (client's and user's).

There is a glitch when the form contains rows of control.
Appending a help-block with the error messages, lead to controls
misplacement. Perhaps error messages should be moved into tooltips
of relative controllers?