Client Side Directives Experiment
In order to remove code duplication from controllers that deal with
validation errors, this experiment introduces a simple directive
to be used on forms with validations.

The directive relies on the fact that the controller in charge for
the page, has an "Errors" property which gets populated at some
point with validation errors from backend.

When that property changes, previous validation errors get cleared
by the directive, and new errors are shown by adding appropriate
classes and/or DOM elements to the UI.

Next step in this experiment is to remove direct dependency on
$scope model by means of $scope variable names. Ad-hoc attribute
should be used.