Client Side Fix to: Users, Clients, Locations form
Fixed: Form submit not enabled if validation errors are present.

The issue:
When a field was marked as "required", if the user clears the value from
the input field, the corresponding field gets removed from the model.
Thus, upon form submit, the attributes passed to the backend for
update/creation are orphan of those fields marked as required and left
empty.

The solution:
Preventing form submission when it's not valid works well.