Server Side Contract information
Since each client can have, at most, one contract, it was a waste of
resources and development efforts, adding Contracts as a nested
resource and dealing with the code required to implement that feature.

Later on, should the need arise for clients to have multiple contracts,
we can always migrate to nested resources.