Client Side Interventions UI
There's an issue with intervention editing/creation form.
The ng-include clause used to clean up the index.html, creates a new
scope which, although inheriting from its parent, creates copies of
primitive properties. This represents a problem since changes in UI
happen on the child scope and are not applied to the parent.

An alternative to ng-include is under research.