Server Side Interventions
Adding specs for Interventions#show
Fixed Location factory, which not included the associated client.
Fixed Location model which did not enforced the presence of an
associated client.

Added RABL partials to Locations and Clients which are used by
Interventions.
The same partials could and should be used by related views.

Added Interventions#show route and method to the backend.

The specs for Interventions are now failing because at some point,
we are testing for what's returned from the backend.
Obviously this has now changed since the inclusion of: client, location
and user in the backend's reply.