Server Side Interventions' Specs
Fixed specs.

Interventions#index now checks for the presence of associations in the
response.

Since RABL templates now use inheritance and have been refactored, it is
possible from now on to rely on incremental specs for response contents.

For example, knowing that the specs for the show action are correct, and
being aware that index action iherits from show action template, we can
safely assume that the required attributes are present in both
responses.

The same reasoning applies to specs of models with associations.