Client Side Interventions Management
Working on adding new interventions.
When adding a new intervention the $scope.intervention object doesn't
get initialized properly and this prevents persistence when leaving any
fields to theri default value!
Leaving a field to its default value doesn't add it to the object's
properties and thus upon post the object sent to the backend is a
partial representation of the user's input.
When the partial representation lacks of required fields, the backend
returns errors to the client.