Client Side Interventions Management
Fixed issue with interventions update. Now, after a successful update,
the form is rendered correctly.

The issue was due the backend not returning the updated version of the
model in PUT's response.

Fixed issue with intervention's edit form that prevented the user from
cancelling the operation. The "Annulla" button wasn't bound anymore to a
click event.