Server Side Specs adjustments to deal with authentication
Added two spec helpers:

login_as(user)
and
logout()

The first one stubs ApplicationController#current_user so that it
returns the user passed via parameters.
This allows us to impersonate a given user (with its roles and
everything) during a spec.

Likewise, the second helper, stubs the same method as before except this
time we are always returning a nil. This has the net effect of a logout
since no current_user is defined anymore.