Client Side Bug Fixing
Fixed a bug that initialized main controller multiple time.
The bug spawned because MainCtrl was assigned to 2 different scopes. One in application Layout,
the other on '/' route.

MainCtrl has been renamed to MainNavCtrl and assigned to NavigationBar in application's Layout.
A WelcomeCtrl has been created and assigned to manage '/' route.

MainNavCtrl.currentUser() has been moved to a property on the scope, since it was invoked multiple
time according to digest's updates.

Server Side
Authentication's Draft

Added Api::V1::BaseController and Api::V1::RestrictedController.
BaseController is for those controllers that doesn't need authentication while, RestrictedController
forces authentication on all descendants.

Fixed a bug on SessionsController that allowed a malicious user to logout some other logged in user.
Upon session destroying the server now checks whether the current user is the same of the one whose
session is requested to terminate. When this check fails, the server replies with status 406 (
unacceptable request).

There still is a bug with SessionController. As of now we can't place the controller under
the correct namespace (Api::V1::). When doing so, the create action fails with a NoMethodError
and complains about to_json missing from rabl_responder (or something like that).
FURTHER INVESTIGATION REQUIRED.