Server Side SessionsController Added authentication status check Clients can now call api/v1/sessions/authenticated_user in order to get the current session user if any. When there's no currently authenticated user the reply will be user: ''.
Client Side
MainController and SessionsSvc
MainController now queries backend upon initialization in order to find
if there's any current logged in user.
This prevents a bug in the UI when the user closes the browser window
without doing a logout. In that case, the session on the server doesn't
get destroyed and if the user goes back to the app, the navbar behaved
like there was no current user logged in, while everything else worked
properly since the old session was still valid.